# backend/docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_norman6464_FreStyle_backend_internal_usecase.ExecuteCodeOutput:
    properties:
      exitCode:
        type: integer
      stderr:
        type: string
      stdout:
        type: string
    type: object
...
  internal_handler.codeExecuteRequest:
    properties:
      code:
        type: string
      language:
        type: string
    required:
    - code
    type: object
  internal_handler.cognitoCallbackReq:
    properties:
      code:
        type: string
      invitationToken:
        description: InvitationToken は招待マジックリンク経由の UUID（任意）。指定時は email 検索より優先して照合する。
        type: string
    required:
    - code
    type: object
...
  internal_handler.errorResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
...
  internal_handler.meResponse:
    properties:
      cognitoSub:
        example: abc-123-uuid
        type: string
      companyId:
        example: 1
        type: integer
      createdAt:
        type: string
      displayName:
        example: 山田 太郎
        type: string
      email:
        example: user@example.com
        type: string
      groups:
        example:
        - admin
        items:
          type: string
        type: array
      id:
        example: 42
        type: integer
      isAdmin:
        example: false
        type: boolean
      onboarded:
        example: true
        type: boolean
      role:
        example: trainee
        type: string
      updatedAt:
        type: string
    type: object
  internal_handler.messageResponse:
    properties:
      message:
        example: ログインしました。
        type: string
    type: object
...
info:
  contact:
    name: FreStyle Engineering
    url: https://normanblog.com
  description: |-
    新卒 IT エンジニア 向け 統合 研修 プラットフォーム の REST API。
    Clean Architecture (port = usecase/repository / adapter = adapter/persistence) で 構築。
  termsOfService: https://normanblog.com/terms
  title: FreStyle Backend API
  version: "2.0"
paths:
...
  /auth/cognito/callback:
    post:
      consumes:
      - application/json
      description: Cognito Hosted UI から の callback。 authorization code を access /
        refresh / id token に 交換 し HttpOnly Cookie で 返す。 新規 user は 招待 or Cognito admin
        group 必須。
      parameters:
      - description: Cognito callback (code 必須、 invitationToken 任意)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.cognitoCallbackReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.messageResponse'
        "400":
          description: code 欠落 等
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: token 交換 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: 招待 なし の 新規 user
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "429":
          description: レート制限超過
          headers:
            Retry-After:
              description: '再試行までの秒数 (例: 60)'
              type: string
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Cognito 未 設定 等 の 内部 エラー
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "502":
          description: Cognito 到達 不可
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Cognito callback (認可 コード → token 交換)
      tags:
      - auth
...
  /auth/cognito/refresh-token:
    post:
      description: refresh_token Cookie で access_token を 再 発行 し HttpOnly Cookie に
        セット する。 失敗 (refresh 切れ 等) は 401 で Cookie クリア。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.messageResponse'
        "401":
          description: refresh_token 欠落 / 無効
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "429":
          description: レート制限超過
          headers:
            Retry-After:
              description: '再試行までの秒数 (例: 60)'
              type: string
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "502":
          description: Cognito 到達 不可
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: アクセス トークン リフレッシュ
      tags:
      - auth
  /auth/me:
    get:
      description: Cookie 認証 を 元 に 現在 ログイン 中 の user 情報 (id / email / role / isAdmin
        / onboarded 等) を 返す。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.meResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: DB に user が ない (Cognito 側 だけ 存在)
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: DB / repository 取得 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: current user 情報 取得
      tags:
      - auth
  /code/execute:
    post:
      consumes:
      - application/json
      description: trainee が 書いた コード を サーバ 側 sandbox で 実行 し stdout/stderr/exitCode
        を 返す。 language は php / go 等。
      parameters:
      - description: コード + 言語
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.codeExecuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_usecase.ExecuteCodeOutput'
        "400":
          description: バリデーション or 実行 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: コード サンドボックス 実行
      tags:
      - code-execution
...
securityDefinitions:
  CookieAuth:
    description: Cognito 由来 の JWT を HttpOnly Cookie で 送る。 ログイン 後 自動 付与。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"